make it work with new rocketh
