+++ pnpm-workspace.yaml
@@ -1,0 +1,3 @@
+overrides:
+  '@rocketh/proxy': link:../rocketh/packages/rocketh-proxy
+  rocketh: link:../rocketh/packages/rocketh
