+++ rocketh/config.ts
@@ -1,0 +1,22 @@
+// ----------------------------------------------------------------------------
+// Typed Config
+// ----------------------------------------------------------------------------
+import type {UserConfig} from 'rocketh';
+
+// this one provide a protocol supporting private key as account
+import {privateKey} from '@rocketh/signer';
+
+export const config = {
+    accounts: {
+        deployer: {
+            default: 0,
+        },
+        admin: {
+            default: 1,
+        },
+    },
+    data: {},
+    signerProtocols: {
+        privateKey,
+    },
+} as const satisfies UserConfig;
